Implement database logging for outError and outCrash. Some error message types will probably need to be tweaked eventually.

+++ src/server/game/Server/WorldSession.cpp
@@ -471,7 +471,7 @@ void WorldSession::LogoutPlayer(bool Save)
         if (Save)
         {
             uint32 eslot;
-            for (int j = BUYBACK_SLOT_START; j < BUYBACK_SLOT_END; ++j)
+            for (uint8 j = BUYBACK_SLOT_START; j < BUYBACK_SLOT_END; ++j)
             {
                 eslot = j - BUYBACK_SLOT_START;
                 _player->SetUInt64Value(PLAYER_FIELD_VENDORBUYBACK_SLOT_1 + (eslot * 2), 0);
